Ditch celery for rq
