+++ Makefile
@@ -1,13 +1,10 @@
-.PHONY: init clean celery assets server db
+.PHONY: init clean assets db
 
 init:
 	pip install -r requirements.txt
 
 clean:
 	find . -name '*.pyc' -delete
-
-celery:
-	python runcelery.py -A app.tasks worker -l INFO
 
 assets:
 	cd app/static && bower install && cd ..
