+++ app/__init__.py
@@ -8,7 +8,7 @@ from app.assets import assets
 from app.auth import auth
 from app.commands import create_db, drop_db, populate_db, recreate_db
 from app.database import db
-from app.extensions import lm, api, travis, mail, bcrypt, celery, babel
+from app.extensions import lm, api, travis, mail, bcrypt, babel, rq
 from app.user import user
 import app.utils as utils
 
@@ -57,9 +57,9 @@ def register_extensions(app):
     lm.init_app(app)
     mail.init_app(app)
     bcrypt.init_app(app)
-    celery.config_from_object(app.config)
     assets.init_app(app)
     babel.init_app(app)
+    rq.init_app(app)
 
 
 def register_blueprints(app):
