+++ app/auth/views.py
@@ -5,7 +5,7 @@ from flask_babel import gettext
 from flask_login import login_user, login_required, logout_user
 from itsdangerous import URLSafeSerializer, BadSignature
 from app.extensions import lm
-from app.tasks import send_registration_email
+from app.jobs import send_registration_email
 from app.user.models import User
 from app.user.forms import RegisterUserForm
 from .forms import LoginForm
@@ -57,7 +57,7 @@ def register():
         s = URLSafeSerializer(current_app.secret_key)
         token = s.dumps(user.id)
 
-        send_registration_email.delay(user.id, token)
+        send_registration_email.queue(user.id, token)
 
         flash(
             gettext(
