+++ app/extensions.py
@@ -1,26 +1,19 @@
-from flask_login import LoginManager
-lm = LoginManager()
-
-from flask_restless import APIManager
-api = APIManager()
-
-from flask_travis import Travis
-travis = Travis()
-
-from flask_mail import Mail
-mail = Mail()
-
-from werkzeug.contrib.cache import SimpleCache
-cache = SimpleCache()
-
-from flask_bcrypt import Bcrypt
-bcrypt = Bcrypt()
-
-from celery import Celery
-celery = Celery()
-
 from flask_assets import Environment
-assets = Environment()
-
 from flask_babel import Babel
+from flask_bcrypt import Bcrypt
+from flask_login import LoginManager
+from flask_mail import Mail
+from flask_restless import APIManager
+from flask_rq2 import RQ
+from flask_travis import Travis
+from werkzeug.contrib.cache import SimpleCache
+
+api = APIManager()
+assets = Environment()
 babel = Babel()
+bcrypt = Bcrypt()
+cache = SimpleCache()
+lm = LoginManager()
+mail = Mail()
+rq = RQ()
+travis = Travis()
